Gibbs sample for tausq
This is the main modification to get the random intercept parameters of
b_i and tausq. The estimated parameters are likely not consistent with
the truth since I have not been able to test the codes in R. However, we
have a developed a method that produces consistent estimates by using
bart from BayesTree to draw the regression trees and then drawing the
sigsqs, tausqs, and bis outside of bart. Hence, I should be able to make
the correct changes once I know how to get modifications on bartMachine
to work.